pyproject.toml
[tool.poetry]
name = "homeworks30"
version = "0.1.0"
description = ""
authors = ["AnastasiaGoryacheva <nastyaitaly@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
